pyproject.toml
[project]
name = "council"
version = "0.1.0"
description = "Council of LLM CLIs — query multiple AI coding agents and synthesize bias-free answers"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.15",
    "rich>=13.0",
    "pyyaml>=6.0",
    "libtmux>=0.40",
]

[project.scripts]
council = "council.cli:app"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
